Eric/ZHOUDIU/TelegramBot/requirements.txt
leancloud>=2.0.0,<3.0.0
Flask>=0.10.1,<1.0.0
python-telegram-bot
